Adds jsonPath to error messages, so errors can be pinpointed. (#53)
* Adds jsonPath to error messages, so errors can be pinpointed.

The jsonPath can be used to provide better error messages to users trying
to create valid jsonLogic expressions

Also:
Fixes some exception messages to use the right operator. ("none", "or").
Aligns behavior to reference implementation for sums of unparseable strings,
as well as for between comparisons of > and >=, and for comparison operators
with arrays of more than three values.

* Handles edge cases of + and * the way jsonlogic.com does

When presented with an array parameter, jsonlogic.com uses the first
value for + and * expressions.

* Uses fixtures for the error tests.

Also fixes a bug in the jsonPath calculation for and/or.

* Bump version to 1.1.0-SNAPSHOT